Cloudex (cloudex v1.4.1)

Cloudex takes care of uploading image files or urls to Cloudinary

Types

upload_result()

Specs

upload_result() ::
  {:ok, Cloudex.UploadedImage.t()}
  | {:error, any()}
  | [ok: Cloudex.UploadedImage.t(), error: any()]

Functions

delete(item_list)

Specs

delete([String.t()]) :: :ok

Delete a list of images

delete(item, opts \\ %{})

Delete an image

delete_prefix(prefix)

Deletes a prefix

start(settings)

Specs

start(settings :: map()) :: {:ok, pid()}

You can start the GenServer that holds the cloudinary api settings by hand using this function.

example

start %{api_key: "key", secret: "s3cr3t", cloud_name: "heaven"}

upload(list, options \\ %{})

Specs

upload(list() | [String.t()], map()) :: upload_result()

Uploads a (list of) image file(s) and/or url(s) to cloudinary