cloudex v0.2.0 Cloudex

Cloudex takes care of uploading image files or urls to Cloudinary

Functions

delete(item_list)

delete([String.t]) :: :ok

Delete an image

start(settings)

start(settings :: map) :: {:ok, pid}

You can start the GenServer that holds the cloudinary api settings by hand using this function.

example

start %{api_key: "key", secret: "s3cr3t", cloud_name: "heaven"}

upload(list, opts \\ %{})

upload(list :: [String.t], %{tags: [String.t]}) :: [%Cloudex.UploadedImage{bytes: term, created_at: term, etag: term, format: term, height: term, original_filename: term, public_id: term, resource_type: term, secure_url: term, signature: term, source: term, tags: term, type: term, url: term, version: term, width: term}]

Uploads a (list of) image file(s) and/or url(s) to cloudinary