cloudex v1.2.3 Cloudex

Cloudex takes care of uploading image files or urls to Cloudinary

Types

upload_result()

upload_result() ::
  {:ok, Cloudex.UploadedImage.t()}
  | {:error, any()}
  | [ok: Cloudex.UploadedImage.t(), error: any()]

Functions

delete(item_list)

delete([String.t()]) :: :ok

Delete an image

start(settings)

start(settings :: map()) :: {:ok, pid()}

You can start the GenServer that holds the cloudinary api settings by hand using this function.

example

start %{api_key: "key", secret: "s3cr3t", cloud_name: "heaven"}

upload(list, options \\ %{})

upload(list() | [String.t()], map()) :: upload_result()

Uploads a (list of) image file(s) and/or url(s) to cloudinary