cloudex v0.1.12 Cloudex

Cloudex takes care of uploading image files or urls to Cloudinary

Functions

start(settings)

Specs

start(settings :: Map.t) :: {:ok, PID.t}

You can start the GenServer that holds the cloudinary api settings by hand using this function.

example

start %{api_key: "key", secret: "s3cr3t", cloud_name: "heaven"}

upload(list, opts \\ %{})

Specs

upload(list :: [String.t], %{tags: [String.t]}) :: [Cloudex.UploadedImage.t]

Uploads a (list of) image file(s) and/or url(s) to cloudinary