cloudex v1.1.3 Cloudex.CloudinaryApi

The live API implementation for Cloudinary uploading

Functions

delete(item)

delete(String.t()) ::
  {:ok, %Cloudex.DeletedImage{public_id: term()}} | {:error, any()}

Deletes an image given a public id

json_result_to_struct(result, source)

json_result_to_struct(map(), String.t()) :: %Cloudex.UploadedImage{
  bytes: term(),
  created_at: term(),
  etag: term(),
  format: term(),
  height: term(),
  moderation: term(),
  original_filename: term(),
  phash: term(),
  public_id: term(),
  resource_type: term(),
  secure_url: term(),
  signature: term(),
  source: term(),
  tags: term(),
  type: term(),
  url: term(),
  version: term(),
  width: term()
}

Converts the json result from cloudinary to a %UploadedImage{} struct

upload(item, opts)

upload(String.t() | {:ok, String.t()}, map()) ::
  {:ok, Cloudex.UploadedImage.t()} | {:error, any()}

Upload either a file or url to cloudinary returns {:ok, %UploadedFile{}} containing all the information from cloudinary or {:error, “reason”}