Cloudex.CloudinaryApi (cloudex v1.4.1)

The live API implementation for Cloudinary uploading

Functions

delete(item, opts \\ %{})

Specs

delete(String.t(), map()) ::
  {:ok, %Cloudex.DeletedImage{public_id: term(), response: term()}}
  | {:error, any()}

Deletes an image given a public id

delete_prefix(prefix, opts \\ %{})

Specs

delete_prefix(String.t(), map()) :: {:ok, String.t()} | {:error, any()}

Deletes images given their prefix

json_result_to_struct(result, source)

Specs

json_result_to_struct(map(), String.t()) :: %Cloudex.UploadedImage{
  bytes: term(),
  context: term(),
  created_at: term(),
  etag: term(),
  format: term(),
  height: term(),
  moderation: term(),
  original_filename: term(),
  phash: term(),
  public_id: term(),
  resource_type: term(),
  secure_url: term(),
  signature: term(),
  source: term(),
  tags: term(),
  type: term(),
  url: term(),
  version: term(),
  width: term()
}

Converts the json result from cloudinary to a %UploadedImage{} struct

upload(item, opts \\ %{})

Specs

upload(String.t() | {:ok, String.t()}, map()) ::
  {:ok, Cloudex.UploadedImage.t()} | {:error, any()}

Upload either a file or url to cloudinary opts can contain: %{resource_type: "video"} which will cause a video upload to occur. returns {:ok, %UploadedFile{}} containing all the information from cloudinary or {:error, "reason"}

url(resource_type)