cloudex v0.1.14 Cloudex.EnvOptions

A simple module to get System Environment Variables (ENV VARS)

Functions

merge(options)

Specs

merge(options :: Map.t) :: Map.t

Pass a Map to merge system variables with. Uses the keys of the map to match the system vars with

Examples

Cloudex.EnvOptions.merge %{my_env_key: nil}
> %{my_env_key: "value from env"}

merge_missing_settings(options)

Specs

merge_missing_settings(options :: Map.t) :: Map.t

Same as merge, but only merges settings which have a nil value in the given map